extract walk into its own pair of files
